requirements.txt
django == 4.2.7
djangorestframework == 3.14.0
psycopg2-binary == 2.9.9
django_filter== 23.4
pytest-django == 4.7.0
model_bakery == 1.17.0
pytest-cov == 4.1.0
